[add] Offline First 도입 5 - id type 변경 및 Supa DB 추가 로직 변경

+++ core/data/src/main/java/com/blue/data/repo/database/TodoRepo.kt
@@ -7,7 +7,7 @@ import kotlinx.coroutines.flow.Flow
 interface TodoRepo : Syncable {
     suspend fun insertData(data: Todo)
     fun readAllData(): Flow<List<Todo>>
-    suspend fun deleteData(id: Int)
-    suspend fun changeCheckBox(id: Int, status: Boolean)
+    suspend fun deleteData(id: Long)
+    suspend fun changeCheckBox(id: Long, status: Boolean)
     fun readSelectedData(date: String): Flow<List<Todo>>
 }
