+++ core/data/src/main/java/com/blue/data/repo/supabase/SupabaseRepo.kt
@@ -9,7 +9,7 @@ interface SupabaseRepo : Syncable {
     fun getAuth(): ComposeAuth
     fun getToken(): String?
     suspend fun insertTodo(data: Todo)
-    suspend fun deleteTodo(id: Int)
+    suspend fun deleteTodo(id: Long)
     suspend fun insertMandalart(data: Mandalart)
-    suspend fun deleteMandalart(id: Int)
+    suspend fun deleteMandalart(id: Long)
 }
